+++ src/main/java/dinkplugin/util/SerializedOffer.java
@@ -14,11 +14,11 @@ public class SerializedOffer {
     private GrandExchangeOfferState state;
     private @SerializedName("itemId") int id;
     private @SerializedName("totalQuantity") int quantity;
-    private int price;
-    private int spent;
+    private long price;
+    private long spent;
 
     public boolean equalsOffer(@NotNull GrandExchangeOffer o) {
         return state == o.getState() && id == o.getItemId() && quantity == o.getTotalQuantity()
-            && price == o.getPrice() && spent == o.getSpent();
+            && price == (long) o.getPrice() && spent == (long) o.getSpent();
     }
 }
